rest controllers for classes and students, modified one-to-many and inheritance model relationships, remove unnecessary constructors, added cors configuration

+++ assistant-web-backend/src/main/java/com/pasionatii/assistant/entity/User.java
@@ -2,15 +2,13 @@ package com.pasionatii.assistant.entity;
 
 import jakarta.persistence.*;
 
-
-
 @Entity
 @Table(name = "user")
 @Inheritance(strategy = InheritanceType.JOINED)
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String email;
     private String password;
@@ -19,11 +17,11 @@ public class User {
     public User() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
